Added a gazebo plugin to ur5.urdf
Change your urdf file with this one.
As the creator of the plugin says: "This is a gazebo plugin that publishes models in the world as a moveit planning scene message".
It gives the information of the objects in the gazebo world to moveit, so that when moveit plans a movement it does not make a plan that goes through the objects (CHAIR). Sometimes it fails and some collisions happen but it never tried to go through the middle of the table like before.