Added new functionality.
The new movement is shifting of a given axis.  For shifting x -> 0 , y -> 1, z-> 2. Value is how much you want to move the robot in the given axis. Basically, it first takes the current position and adds the value to the given axis.
Using shift does not change the orientation of the end-effector. Meaning that it does not try to align its end-effector frame with the world frame.